Add files via upload
 I completed the Spring Boot task to perform select, insert, update and delete operations on the task list and employee task list.
The task consists of six APIs to perform CRUD operations on two tables in the database. Every API initially requires authorization from the user. The four out of six Api's are to
perform insert, update, delete, and select operations on the tasklist table respectively. One API is to perform an insert operation on the employee tasklist table using 
created.
The remaining API is to update the status of two tables. For every input field, I prepared a validation according to the requirement; for example, the date should be in the 
of DD-MM-YYYY and more.