Add files via upload
 the Spring Boot task to schedule meetings for the employee, hr and project manager. Here, the employee does not have access to invite HR or the project manager, and the project manager and project manager have access to schedule meetings for the employee and validate the user input data, logins
based on their respective user name, user password and internally project ID, employee ID, generated meeting id, add respective employee id's to database, meeting status whether created or completed and it checks whether the meeting already created or not, if already created it will display message and user name employee who created and if not created the meeting then the data will be inserted into the database. This task is an extension of the previous task.