Add files via upload
We have to enable the Gmail API from the Google Cloud APIs and create a project in Google Cloud. Now, in the credentials section, click on "Help Me To Choose" and fill in the respective details. After filling out the details, you will get a JSON file. Download it and place the JSON file in the src/main/resources directory. Now you can access the Gmail API services to send emails to others. After successfully sending the mail, a tokens folder will be created. We should remove the token folder every time we login with new mail; otherwise, the API will take old email that was previously logged in.