Add files via upload
We have to enable the Gmail API from the Google Cloud APIs by searching "Gmail Api" and clicking enable. Create a project in Google Cloud. In the credentials section, click on "Help Me to Choose" and fill in the respective details. After filling out the details, you will finally get a JSON file.
Note: The downloaded JSON file is not in this repository. Follow the below instructions and place the JSON file in the respective directory.
Download it and place the JSON file in the src/main/resources directory. Now you can access the Gmail API services to send emails to others. After successfully sending the mail, a tokens folder will be created. We should remove the token folder every time we login with new mail; otherwise, the API will take old email that was previously logged in. I have taken the respective teammates employee ID and checked against the respective ID in the database using PROJECT ID. Now, after inserting the accurate employee meeting id into the meeting table in the database, the accurate employee id's email subject and email body are passed into the GMAIL API as parameters. Now, validate the current employee through POP WINDOW on WEB BROWSER after completion of validation on the web. Split the employee string array, and for each employee ID in the array, get the respective email ID from the employee table in the database and send a meeting email to all the employees of their respective projects.